refactor(home): decompose HomePageInner god component into section components (#128)
Extract each landing section from App.tsx into src/components/home/:
HeroSection, PainSection, SocialProofSection, VideoSection, PlanSection,
CatalogSection (owns search/category filtering), and FounderSection,
composed by HomePage. getRandomPrompt moves to lib/randomPrompt and is
re-exported from App so existing characterization tests hold unchanged.
