+++ src/lib/randomPrompt.ts
@@ -1,0 +1,7 @@
+import designs from '@/data/designs';
+import { buildAgentPrompt } from '@/lib/agentPrompt';
+
+export function getRandomPrompt(rng: () => number = Math.random) {
+  const pick = designs[Math.floor(rng() * designs.length)];
+  return buildAgentPrompt(pick.jsonUrl);
+}
